+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Unreleased
 - Drop Chef 11 compatibility, test against Chef 13 and 12
 - Convert all the HWRPs to Chef custom resources
 - Patch the custom resources for Chef 13 compatibility
+- Make the config resource's path configurable
 
 v1.1.0 (2016-08-11)
 -------------------